requirements2.txt
absl-py==0.9.0
astroid==2.0.4
atomicwrites==1.4.0
attrs==20.3.0
cachetools==4.1.1
carla==0.9.6
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.1
grpcio==1.36.1
gym==0.17.1
idna==2.10
imageio==2.8.0
importlib-metadata==2.0.0
isort==4.3.4
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.2.2
matplotlib==3.0.3
mccabe==0.6.1
more-itertools==8.6.0
networkx==2.4
numpy==1.14.2
oauthlib==3.1.0
olefile==0.46
pandas==0.25.3
pathlib2==2.3.2
Pillow==5.2.0
pluggy==0.11.0
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pyglet==1.5.0
pylint==2.1.1
pyparsing==2.4.7
pytest==3.8.1
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.25.0
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.6
scikit-image==0.15.0
scipy==1.4.1
seaborn==0.9.1
tabulate==0.8.7
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.53.0
transforms3d==0.3.1
typed-ast==1.1.0
urllib3==1.26.2
wget==3.2
wrapt==1.10.11
zipp==1.2.0
